moved the navbar to its own componentfolder and fixed imports
